REF renamed file validation trait
