+++ CommonLogic/Filesystem/InMemoryFile.php
@@ -349,6 +349,9 @@ class InMemoryFile implements FileInterface, FileInfoInterface
      */
     public function getMimetype(): ?string
     {
+        if ($this->mimeType === null && null !== $binary = $this->read()) {
+            $this->mimeType = MimeTypeDataType::findMimeTypeOfString($binary);
+        }
         if ($this->mimeType === null) {
             $this->mimeType = MimeTypeDataType::guessMimeTypeOfExtension($this->getExtension(), 'text/plain');
         }
